# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Project Laravel",
        "description": "Project Laravel",
        "contact": {
            "email": "bertrand@idearobin.com"
        },
        "version": "1.0.0"
    },
    "host": "api.canard.dev",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieves all Users",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "data": {
                                    "description": "Get all Users.",
                                    "type": "array"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Could not retrieve users",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "error": {
                                    "description": "Get all Users.",
                                    "type": "string",
                                    "default": "could_not_retrieve_data"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/users/profile/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieve single User details",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "USER_ID to retrieve",
                        "required": true,
                        "type": "integer",
                        "default": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "data": {
                                    "description": "Get user details.",
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Could not retrieve data",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "error": {
                                    "description": "Get user details.",
                                    "type": "string",
                                    "default": "could_not_retrieve_data"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete User",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "USER_ID to delete",
                        "required": true,
                        "type": "integer",
                        "default": 8
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "status": {
                                    "description": "Status message from server",
                                    "type": "string",
                                    "default": "Record successfully deleted."
                                },
                                "user": {
                                    "description": "Delete User.",
                                    "$ref": "#/definitions/User"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Could not delete data",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "error": {
                                    "description": "Delete User.",
                                    "type": "string",
                                    "default": "could_not_delete_data"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Edit User",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "USER_ID to edit",
                        "required": true,
                        "type": "integer",
                        "default": 8
                    },
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User object to edit",
                        "required": true,
                        "schema": {
                            "title": "user",
                            "properties": {
                                "user": {
                                    "description": "Edit User.",
                                    "$ref": "#/definitions/User"
                                }
                            },
                            "type": "object"
                        },
                        "type": "object"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "title": "data",
                            "required": [
                                "status",
                                "user"
                            ],
                            "properties": {
                                "status": {
                                    "description": "Status message from server",
                                    "type": "string",
                                    "default": "Record successfully updated."
                                },
                                "user": {
                                    "description": "Edit User.",
                                    "$ref": "#/definitions/User"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Could not update data",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "error": {
                                    "description": "Edit User.",
                                    "type": "string",
                                    "default": "could_not_update_data"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Retrieves User Profile",
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "data": {
                                    "description": "Get logged in user profile.",
                                    "type": "array"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Could not retrieve data",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "error": {
                                    "description": "Get logged in user profile.",
                                    "type": "string",
                                    "default": "could_not_retrieve_data"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add User",
                "parameters": [
                    {
                        "name": "user",
                        "in": "body",
                        "description": "User object to add",
                        "required": true,
                        "schema": {
                            "title": "user",
                            "properties": {
                                "user": {
                                    "description": "Add User.",
                                    "$ref": "#/definitions/User"
                                }
                            },
                            "type": "object"
                        },
                        "type": "object"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "title": "data",
                            "required": [
                                "status",
                                "user"
                            ],
                            "properties": {
                                "status": {
                                    "description": "Status message from server",
                                    "type": "string",
                                    "default": "Record successfully added."
                                },
                                "user": {
                                    "description": "Add User.",
                                    "$ref": "#/definitions/User"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Could not add data",
                        "schema": {
                            "title": "data",
                            "properties": {
                                "error": {
                                    "description": "Add User.",
                                    "type": "string",
                                    "default": "could_not_add_data"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "properties": {
                "email": {
                    "description": "Unique email address",
                    "type": "string",
                    "default": "janedoe@example.com"
                },
                "password": {
                    "description": "Password must be at least 6 characters",
                    "type": "string",
                    "default": "helloworld"
                },
                "password_confirmation": {
                    "description": "Re-enter password",
                    "type": "string",
                    "default": "helloworld"
                },
                "first_name": {
                    "description": "First Name",
                    "type": "string",
                    "default": "Jane"
                },
                "last_name": {
                    "description": "Last Name",
                    "type": "string",
                    "default": "Doe"
                },
                "phone": {
                    "description": "Phone number",
                    "type": "string",
                    "default": "555-1234"
                },
                "address": {
                    "description": "Address",
                    "type": "string",
                    "default": "Galway St."
                },
                "suburb": {
                    "description": "Suburb",
                    "type": "string",
                    "default": "Adamantine"
                },
                "postcode": {
                    "description": "Postal Code",
                    "type": "string",
                    "default": "90218"
                },
                "state_id": {
                    "description": "STATE_ID of state",
                    "type": "integer",
                    "format": "int64",
                    "default": 2
                }
            }
        }
    },
...
}